-> Implemented Spring API Gateway.

+++ SplitwiseUserService/src/main/java/com/splitwise/userservice/services/Impl/GroupServiceImpl.java
@@ -4,19 +4,22 @@ import com.splitwise.userservice.entities.Group;
 import com.splitwise.userservice.entities.User;
 import com.splitwise.userservice.exceptions.ResourceNotFound;
 import com.splitwise.userservice.payload.APIResponse;
-import com.splitwise.userservice.payload.ExitGroupBody;
+import com.splitwise.userservice.payload.UserGroupBody;
 import com.splitwise.userservice.payload.Expense;
 import com.splitwise.userservice.payload.GroupDetail;
-import com.splitwise.userservice.payload.UserListResponse;
 import com.splitwise.userservice.repositories.GroupRepository;
 import com.splitwise.userservice.repositories.UserRepository;
 import com.splitwise.userservice.services.GroupService;
+import com.splitwise.userservice.services.externalServices.ExpenseService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 @Service
@@ -30,6 +33,10 @@ public class GroupServiceImpl implements GroupService {
 
     @Autowired
     RestTemplate restTemplate;
+
+    @Autowired
+    private ExpenseService expenseService;
+
     @Override
     public Group createGroup(Group group) {
         User user = this.userRepository.findById(group.getOwnerID()).orElseThrow(() -> new ResourceNotFound("User", "Id"));
@@ -61,13 +68,48 @@ public class GroupServiceImpl implements GroupService {
 
     }
 
+    @Transactional
     @Override
-    public void deleteGroup(String groupID) {
-        this.groupRepository.deleteById(groupID);
+    public APIResponse deleteGroup(String groupID, String ownerID) {
+        Group group = this.groupRepository.findById(groupID).orElseThrow(() -> new ResourceNotFound("Group", "Id"));
+
+        APIResponse apiResponse = new APIResponse();
+        try {
+            if(group.getOwnerID().compareTo(ownerID) == 0) {
+
+
+                // if group is successfully deleted, we can delete expenses for this group in expense service
+                try {
+                    APIResponse response = expenseService.deleteExpensesOfGroup(groupID);
+                    if(response.getSuccess()) {
+                        apiResponse.setObject("Group and expenses deleted!");
+                        apiResponse.setSuccess(true);
+                    } else {
+                        apiResponse.setObject("Error in deleting expenses");
+                        apiResponse.setSuccess(false);
+                    }
+                } catch (Exception e) {
+                    apiResponse.setObject("error in deleting expenses");
+                    apiResponse.setSuccess(false);
+                    return apiResponse;
+                }
+
+                // delete group
+                this.groupRepository.deleteById(groupID);
+            } else {
+                apiResponse.setObject("Only owner can delete a group!");
+                apiResponse.setSuccess(false);
+            }
+            return apiResponse;
+        } catch(Exception e) {
+            apiResponse.setObject("Error, deleting group!");
+            apiResponse.setSuccess(false);
+            return apiResponse;
+        }
     }
 
     @Override
-    public APIResponse exitGroup(ExitGroupBody exitGroupBody) {
+    public APIResponse exitGroup(UserGroupBody exitGroupBody) {
         try {
             Group group = this.groupRepository.findById(exitGroupBody.getGroupID()).orElseThrow(() -> new ResourceNotFound("Group", "Id"));
             Set<User> userList = group.getUserList();
@@ -104,7 +146,7 @@ public class GroupServiceImpl implements GroupService {
 
         // get expense list of group from expense service
 
-        Set<Expense> expenses = restTemplate.getForObject("http://EXPENSE-SERVICE/api/expense/get-expense-list-by-group-id/"+group.getGroupID(), HashSet.class);
+        Set<Expense> expenses = expenseService.getExpenses(group.getGroupID());
         groupDetail.setExpenseList(expenses);
 
         return new APIResponse(groupDetail,true);
